move VertoHandler.spec file to the same folder of VertoHandler to have automatic jest mocking
